@@ -1,16 +1,16 @@
 module github.com/aerokube/ggr
 
-go 1.18
+go 1.19
 
 require (
-	github.com/aandryashin/matchers v0.0.0-20160729131923-5eb67beb188b
+	github.com/aandryashin/matchers v0.0.0-20161126170413-435295ea180e
 	github.com/aandryashin/reloader v0.0.0-20161127125235-da4f1b43ce40
 	github.com/abbot/go-http-auth v0.4.1-0.20220112235402-e1cee1c72f2f
-	golang.org/x/net v0.0.0-20210226172049-e18ecbb05110
+	golang.org/x/net v0.2.0
 )
 
 require (
 	github.com/fsnotify/fsnotify v1.5.4 // indirect
 	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519 // indirect
-	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
+	golang.org/x/sys v0.2.0 // indirect
 )
